Refactoring FileIO methods to be used for saving and laoding to Database
the DAO classes have been made to refactor much of the fuctionality if the reading and writting of the fileIo class, so that saving and loading can be done through the database instead. Also fromData methods were created for the classes, move, Playerprofile, LeaderBoard and GameState, so that data from the database tables can be  converted into the objects.