Basic MainMenu
the gui asks for player name and loads the player, the gui displays the main menu after player is loaded. The game manager method new player is updated, and code is commented out of the userinterface class, Gamecontroller class, Gameapp class, GameMangerclass. The commented out code is from the CLI and will be removed later.