GUI can now load the leaderboard and create or load player profiles when their names are entered
the the methods in MainMenuGUI:
selectGameModeAndSize() and displayLeaderboard()  are made so that the gui can load games and display the leaderboard.
showLeaderboardForDifficulty class in Gamemanger class, is modified to return a string, which can then be displayed in the gui. The newplayer method is also modified to work with the gui