Integrate GUI with game logic
Main changes:
GameManager class methods modified:
 newgame():
 initilizeds the game, calls selectgamemode from the mainmenugui class to set up the game, creates new BoardGUI object and returns that object.
loadgame method,  loads the plaeyes selected gamestate, execute the saved moves,creates  a new boardgui object and returns it.
GameEventlistener:The GameEventListener interface implements the Observer Pattern", it is used in the Gamecontroller class, to observe any game ending events, win or lose.

The gamecontroller class is modified, to call the mainmengui instead of the old userinterface class, and set it as visible, the startnewgame and loadnewgame methods are use to call the newgame and load game methods from the gamemanger class.